use float instead of double in C data structure to further save the space. #79

+++ src/GatingHierarchy.cpp
@@ -344,7 +344,7 @@ void GatingHierarchy::transforming(bool updateCDF)
 			if(curTrans->gateOnly())
 				continue;
 
-			valarray<double> x(this->fdata.subset(curChannel));
+			valarray<float> x(this->fdata.subset(curChannel));
 			if(dMode>=GATING_HIERARCHY_LEVEL)
 				cout<<"transforming "<<curChannel<<" with func:"<<curTrans->getChannel()<<endl;
